+++ vision/VisualCorner.h
@@ -80,10 +80,11 @@ private: // private methods
     const shape getLClassification();
     void determineCornerIDFromShape();
     inline float cornerDistanceToSD(float _distance) {
-        return sqrt(2.0f * std::max(10 + _distance*0.00125f, 250.0f));
+        return static_cast<float>(sqrt(2.0f * std::max(10.f + _distance*0.00125f,
+                                                       250.0f)));
     }
     inline float cornerBearingToSD(float _bearing) {
-        return sqrt(static_cast<float>(M_PI) / 4.0f);
+        return static_cast<float>(sqrt(M_PI / 4.0f));
     }
